requirements.txt
absl-py==2.1.0
accelerate==0.30.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
apturl==0.5.2
asteroid-filterbanks==0.4.0
attrs==24.2.0
audioread==3.0.1
bcrypt==3.2.0
blinker==1.8.1
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
ci-sdr==0.0.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
ConfigArgParse==1.7
contourpy==1.2.1
cryptography==3.4.8
ctc-segmentation==1.7.4
cupshelpers==1.0
cycler==0.12.1
Cython==3.0.10
dataclasses==0.6
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
Distance==0.1.3
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docker-pycreds==0.4.0
duplicity==0.8.21
editdistance==0.8.1
einops==0.8.0
espnet==202402
espnet-tts-frontend==0.0.3
fairscale==0.4.13
fast-bss-eval==0.1.3
fasteners==0.14.1
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
future==0.18.2
g2p-en==2.1.0
gitdb==4.0.11
GitPython==3.1.43
gunicorn==22.0.0
h5py==3.11.0
httplib2==0.20.2
huggingface-hub==0.22.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.3
importlib-metadata==4.6.4
indic-num2words==1.2.1
indic-unified-parser==1.0.6
inflect==7.0.0
itsdangerous==2.2.0
jaconv==0.3.4
jamo==0.4.1
jax==0.4.34
jaxlib==0.4.34
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.0
kaldiio==2.18.0
keras==3.3.3
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
librosa==0.9.2
llvmlite==0.42.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mediapipe==0.10.15
ml-dtypes==0.4.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
namex==0.0.8
netifaces==0.11.0
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
paramiko==2.9.3
peft==0.11.1
pexpect==4.8.0
Pillow==9.0.1
pipenv==11.9.0
platformdirs==4.2.1
pooch==1.8.1
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
PyAudio==0.2.11
pycairo==1.20.1
pycocotools==2.0.7
pycparser==2.22
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==2.4.7
pypinyin==0.44.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyworld==0.3.4
pyxdg==0.27
PyYAML==5.4.1
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
resampy==0.4.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.13.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
-e git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588#egg=segment_anything
sentencepiece==0.1.97
sentry-sdk==2.1.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sounddevice==0.5.1
soundfile==0.12.1
ssh-import-id==5.11
sympy==1.12
systemd-python==234
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.3.1
torch-complex==0.4.3
torchsummary==1.5.1
torchvision==0.18.1
tqdm==4.66.2
transformers==4.39.3
triton==2.3.1
typeguard==2.13.3
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.8
urllib3==2.2.1
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
wandb==0.17.0
Werkzeug==3.0.2
xdg==5
xkit==0.0.0
zipp==1.0.0
